requirements.txt
# Lima Air Quality Dashboard - Requirements

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# HTTP requests and APIs
requests>=2.31.0
urllib3>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Data validation and cleaning
scikit-learn>=1.3.0
scipy>=1.11.0

# Date and time processing
python-dateutil>=2.8.0
pytz>=2023.3

# Logging and monitoring
structlog>=23.1.0

# Database migrations (optional)
alembic>=1.11.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Future phases dependencies (uncomment when needed)
# 
# # Phase 3: ML and forecasting
# prophet>=1.1.4
# tensorflow>=2.13.0
# torch>=2.0.0
# 
# # Phase 4: API
# fastapi>=0.103.0
# uvicorn>=0.23.0
# pydantic>=2.0.0
# 
# # Phase 5: Dashboard
# streamlit>=1.26.0
# plotly>=5.15.0
# folium>=0.14.0
# pydeck>=0.8.0
# 
# # Phase 6: Notifications
# twilio>=8.6.0
# celery>=5.3.0
# redis>=5.0.0
